refactor: change the repositories to the typeorm repositories
- Change in the category repository
